fix: fix some potentially problematic code (#548)
- fix initialization of union vars, using declaration, assignments to define vars, not forced type conversion
- fix initialization of 'bool is_write' crossing 'goto bad', define 'bool is_write' at first
- fix passing to parameter to function, the 'update_global' func take no arguments
- fix some vars type mismatch
---------

+++ src/engine/interpreter/fp.c
@@ -42,7 +42,7 @@ static inline float16_t rtlToVF16(rtlreg_t r) {
 }
 
 static inline float32_t rtlToVF32(rtlreg_t r) {
-  float32_t f = { .v = r };
+  float32_t f = { .v = (uint32_t)r };
   return f;
 }
